lawyer forgotpassword token expiry time changed and added validation for that and move the app exeption and error in common folder
