+++ Backend/module/auth/lawyerAuth/application/lawyer-use-case/lawyerChangePasswordApplication.ts
@@ -1,5 +1,5 @@
-import { AppError } from "../../../userAuth/application/error/AppEnumError";
-import { AppException } from "../../../userAuth/application/error/errorException";
+import { AppError } from "../../../../../common/error/AppEnumError";
+import { AppException } from "../../../../../common/error/errorException";
 import { IHashService } from "../../../userAuth/infrastructure/services/IhashService";
 import { IChangePasswordRepositorie } from "../../interface/repositories/IchangePasswordRepositorie";
 import { ILawyerChangePasswordApplication } from "../lawyer-use-case-interface/IlawyerChangePasswordApplication";
@@ -14,12 +14,13 @@ export class LawyerChangePasswordApplication implements ILawyerChangePasswordApp
 
     async changePassword(email: string, password: string, token: string): Promise<void> {
         try {
-            console.log(email,password,token)
             let decodeToken:any=jwt.decode(token)
 
             if(Date.now()>decodeToken?.exp*1000){
-                throw new AppException("Link has Expired try again")
+                throw new AppException("Link has Expired try again",410)
             }
+
+            jwt.verify(token,process.env.FORGOT_PASSWORD_TOKEN_SECRET!)
 
             let hashedPassword:string=await this._hashService.hash(password)
