+++ Backend/module/auth/lawyerAuth/application/lawyer-use-case/lawyerForgotPasswordApplication.ts
@@ -1,6 +1,6 @@
-import { AppError } from "../../../userAuth/application/error/AppEnumError";
-import { AppException } from "../../../userAuth/application/error/errorException";
-import { AppStatusCode } from "../../../userAuth/application/statusCode/AppStatusCode";
+import { AppError } from "../../../../../common/error/AppEnumError";
+import { AppException } from "../../../../../common/error/errorException";
+import { AppStatusCode } from "../../../../../common/statusCode/AppStatusCode";
 import { ILawyerSignup } from "../../domain/entity/lawyerEntity";
 import { IForgotPasswordEmailService } from "../../infrastructure/service/IforgotPasswordEmailService";
 import { IForgotPasswordTokenGeneration } from "../../infrastructure/service/IforgotPasswordTokenGeneration";
